+++ R/RcppExports.R
@@ -13,6 +13,10 @@ SSSreconstruction_cpp <- function(pts, scaleIterations, nneighs, nsamples, separ
     .Call(`_cgalMeshes_SSSreconstruction_cpp`, pts, scaleIterations, nneighs, nsamples, separateShells, forceManifold, borderAngle)
 }
 
+gatherVertices <- function(Vertices, Faces) {
+    .Call(`_cgalMeshes_gatherVertices`, Vertices, Faces)
+}
+
 gTriangle <- function(A, B, C, s, iterations) {
     .Call(`_cgalMeshes_gTriangle`, A, B, C, s, iterations)
 }