+++ src/RcppExports.cpp
@@ -53,6 +53,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// gatherVertices
+Rcpp::List gatherVertices(Rcpp::NumericMatrix Vertices, Rcpp::IntegerMatrix Faces);
+RcppExport SEXP _cgalMeshes_gatherVertices(SEXP VerticesSEXP, SEXP FacesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type Vertices(VerticesSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerMatrix >::type Faces(FacesSEXP);
+    rcpp_result_gen = Rcpp::wrap(gatherVertices(Vertices, Faces));
+    return rcpp_result_gen;
+END_RCPP
+}
 // gTriangle
 Rcpp::XPtr<EMesh3> gTriangle(Rcpp::NumericVector A, Rcpp::NumericVector B, Rcpp::NumericVector C, double s, int iterations);
 RcppExport SEXP _cgalMeshes_gTriangle(SEXP ASEXP, SEXP BSEXP, SEXP CSEXP, SEXP sSEXP, SEXP iterationsSEXP) {
@@ -105,6 +117,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_cgalMeshes_AFSreconstruction_cpp", (DL_FUNC) &_cgalMeshes_AFSreconstruction_cpp, 2},
     {"_cgalMeshes_MinkowskiSum_cpp", (DL_FUNC) &_cgalMeshes_MinkowskiSum_cpp, 2},
     {"_cgalMeshes_SSSreconstruction_cpp", (DL_FUNC) &_cgalMeshes_SSSreconstruction_cpp, 7},
+    {"_cgalMeshes_gatherVertices", (DL_FUNC) &_cgalMeshes_gatherVertices, 2},
     {"_cgalMeshes_gTriangle", (DL_FUNC) &_cgalMeshes_gTriangle, 5},
     {"_cgalMeshes_meshTopology", (DL_FUNC) &_cgalMeshes_meshTopology, 4},
     {"_cgalMeshes_sTriangle", (DL_FUNC) &_cgalMeshes_sTriangle, 6},